+++ packages/gui/src/components/schemas/types.ts
@@ -1,0 +1,8 @@
+import { ComponentType } from 'react'
+import { EditorPropsWithLabel } from '../../types/editor'
+
+export interface DataTypeSchema<T> {
+  input: ComponentType<EditorPropsWithLabel<T>>
+  stringify(value: T): string
+  defaultValue: T
+}
